+++ changelog.md
@@ -1,6 +1,9 @@
 # Change Log
 The BounCA change history
 
+## Develop
+* Added Serial and Fingerprint to certificate table, including search
+
 ## [0.3.0] - Release 2022-02-14
 * Added resend e-mail button if user has not verified email yet
 * Form for admin account generation for fresh installs